refactor: split options.go into top-level + per-output (#540)
options.go was 561 lines — over the project's ~300-line guideline.
Split per-output configuration (WithOutputs, WithNamedOutput,
OutputOption, plus the per-output knobs) into a new
options_output.go file. Pure refactor; no behaviour change, no
public API change.

Kept in options.go (402 lines):
- Option func type + godoc
- Top-level auditor options: WithTaxonomy, WithMetrics, WithAppName,
  WithHost, WithTimezone, WithSynchronousDelivery,
  WithDiagnosticLogger, WithStandardFieldDefaults, WithFormatter,
  WithQueueSize, WithShutdownTimeout, WithValidationMode,
  WithOmitEmpty, WithDisabled, WithSanitizer
- Shared internal helpers: buildLabelSet, checkDestinationDup
  (used by both halves)

Moved to options_output.go (190 lines):
- WithOutputs
- OutputOption func type + outputEntryBuilder struct
- Per-output options: WithRoute, WithOutputFormatter,
  WithExcludeLabels, WithHMAC
- WithNamedOutput
- Auditor.addNamedOutput method

options.go remains slightly over the 300-line target; further
splitting would fragment the cluster of top-level options without
clear benefit. The per-output split was the main reduction target.

Verification:
- go build ./... → clean
- go test -race -count=1 . → green (179s)

Refs #540.
